+++ examples/bench/chat_parser.pl
@@ -15,11 +15,9 @@ go:-
     write('execution time is '),write(T), write(milliseconds).
 
 chat_parser :-
-    my_string(X),
-    determinate_say(X,_),
-    fail.
-chat_parser.
-
+	% original code used a failure-driven loop; changed to use unrolled
+	% forall/2 spo that unexpected failures will make chat_parser/0 fail
+	\+ (my_string(X), \+ determinate_say(X,Y)).
 
 %  query set